This branch was intended to refactor threads with the idea that some control sources and sinks would need there own threads while others could share threads, since they depend on each other anyway. This turned out to be far messier and confusing than any advantage might justify.

Since the initial target for this refactoring involved the USB host, that became the focus of the work. A number of issues needed to be resolved around using the TinyUSB stack, in particular with SoftRocks. It turns out that TinyUSB had a bug involving low speed USB (as used by SoftRocks due to their ATTiny85 chip), which was fixed by pulling the latest main into the third-party/tinyusb_src submodule. There was also a problem with move semantics corrupting the static instance of UsbHost.

USBControlSinks was added, initially as a place to manage the threads discussed above, but it is useful to have this class to manage USB sinks anyway. This required adding a corresponding configuration struct and json handler, and moving USB control sinks down into the new collection.

So currently the stm32h745i-cm4/indirect-update example is able to instantiate a RadioControl, configure it (with hard-wired configs) and have rotary encoder clicks routed through to frequency setting packets to a connected SoftRock.

There is a lot of tidying up to do, but this is enough for one branch.

+++ src/settings/control/radio/src/RadioControl.cpp
@@ -138,7 +138,7 @@ RadioControl::start()
     }, pSink);
     printf("[RadioControl]\t...done!\r\n");
     if (rc != ResultCode::OK) {
-      printf("[RadioControl]\t Error starting sink: %ld.\r\n", rc);
+      printf("[RadioControl]\t Error starting sink: %u.\r\n", static_cast<uint32_t>(rc));
       return rc;
     }
   }
